fix(test): i32 -> i16

+++ tests/test_parser.zig
@@ -146,7 +146,7 @@ test "var statement" {
     defer res.deinit(std.testing.allocator);
     try std.testing.expect(res.stmt == .var_stmt);
     try std.testing.expectEqualStrings("b", res.stmt.var_stmt.name);
-    try std.testing.expectEqual(ast.PrimitiveType.i32, res.stmt.var_stmt.type_ann.?.base.primitive);
+    try std.testing.expectEqual(ast.PrimitiveType.i16, res.stmt.var_stmt.type_ann.?.base.primitive);
     try std.testing.expect(res.stmt.var_stmt.value.* == .literal);
     try std.testing.expectEqualStrings("69", res.stmt.var_stmt.value.literal.raw);
 }
@@ -156,7 +156,7 @@ test "const statement" {
     defer res.deinit(std.testing.allocator);
     try std.testing.expect(res.stmt == .const_stmt);
     try std.testing.expectEqualStrings("b", res.stmt.const_stmt.name);
-    try std.testing.expectEqual(ast.PrimitiveType.i32, res.stmt.const_stmt.type_ann.?.base.primitive);
+    try std.testing.expectEqual(ast.PrimitiveType.i16, res.stmt.const_stmt.type_ann.?.base.primitive);
     try std.testing.expect(res.stmt.const_stmt.value.* == .literal);
     try std.testing.expectEqualStrings("69", res.stmt.const_stmt.value.literal.raw);
 }
